src/test/java/com/myStore/testCases/LoginPageTest.java
package com.myStore.testCases;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.DataProvider;
import org.testng.annotations.Parameters;
import org.testng.annotations.Test;

import com.myStore.base.BaseClass;
import com.myStore.dataProvider.DataProviders;
import com.myStore.pageObjects.HomePage;
import com.myStore.pageObjects.IndexPage;
import com.myStore.pageObjects.LoginPage;
import com.myStore.utility.Log;

public class LoginPageTest extends BaseClass {
	IndexPage indexPage;
	LoginPage loginPage;
	HomePage homePage;

	@Parameters("browser")
	@BeforeMethod(groups= {"Smoke","Sanity","Regression"})
	public void setup(String browser) {
		launchApp(browser);
	}

	
@Test(dataProvider = "credentials",dataProviderClass = DataProviders.class,groups = {"Smoke","Sanity"} )
	public void loginTest(String userName,String password) throws Throwable {
		Log.startTestCase("LoginTest");
		indexPage = new IndexPage();
		loginPage = indexPage.clickOnSignIn();
	//	homePage = loginPage.login(prop.getProperty("username"), prop.getProperty("password"));
		homePage = loginPage.login(userName, password);
		System.out.println(userName+" "+password);
		String actualUrl=homePage.getCurrURL();
		String expectedUrl="http://automationpractice.com/index.php?controller=my-account";
		Assert.assertEquals(actualUrl, expectedUrl);

	}

	@AfterMethod(groups= {"Smoke","Sanity","Regression"})
	public void tearDown() {
		getDriver().quit();
	}

}
